New file types, new features
Added file types : cur,dat,ddf,dds,dll, gif, ini, jpg,json, log,
tga,ver, xml
Added features :
- explorer.php : allow browsing and viewing of lol extracted folder
(using dargon or other tool)
- indexer.php : index lol files (very basic) ( +- 40min full Lol on Win7
x64 4x2.6Ghz)
- search.php : search value in indexed files (very basic) (+- 45 s per
search)

+++ riotDecode/luaobj/LuaobjFile.php
@@ -1,37 +1,44 @@
 <?php
 namespace riotDecode\luaobj;
 
-class LuaobjFile {
-    protected $file;
+use riotDecode\common\BaseFile;
+class LuaobjFile extends BaseFile {
+	
 
 	public function __construct($file) {
-		$this->file = $file;
-
-		$tmpFile = fopen(getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj', "w+");
-		fwrite($tmpFile, $file->getContent());
-		fclose($tmpFile);
-
-		$command = getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'luadec.exe "' . getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj' . '"';
-		$result = `$command`;
-
-		echo $result . "<hr />";
-
-		$command = getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'luadec.exe -dg -dis "' . getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj' . '"';
-		$result = `$command`;
-
-		echo $result . "<hr />";
-
-		unlink(getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj');
+		parent::__construct($file);
+		$this->indexable = true;
 	}
 
-	function getValues() {
-		$this->decodeFile();
-
-		return $this->values;
-	}
-
-	public function __tostring() {
-		return '';
-	}
+	public function &getValues($translateKeys = true) {
+    	if ($this->values == null)
+    	{
+    		$tmpFile = fopen(getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj', "w+");
+    		fwrite($tmpFile, $this->loadFileContent());
+    		fclose($tmpFile);
+    		
+    		$command = getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'luadec.exe "' . getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj' . '"';
+    		$result = `$command`;
+    		
+    		$this->values["content"]["Decompiled"]=$result;
+    		
+    		
+    		$command = getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'luadec.exe -dg -dis "' . getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj' . '"';
+    		$result = `$command`;
+    		
+    		$this->values["content"]["Disassembled"]=$result;
+    		
+    		unlink(getcwd() . DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, explode('\\', __NAMESPACE__)) . DIRECTORY_SEPARATOR . '_private' . DIRECTORY_SEPARATOR . 'tmp.luaobj');
+    		
+    		ksort($this->values);
+    	}
+        return $this->values;
+    }
+    
+    public function showValue($value)
+    {
+    	return $value;
+    }
+	
 }
 ?>
